Change `GET api/party/get` mapping to `GET api/party`; Rename `Forest` to `Obstacle`; Move renamed `WorldAdminService.determinePlaceType` to `fromPlaceBase` and moved it over to `PlaceType`, adding an associated registry in the form of a static type map; Do the same for `CreatureType.fromCreatureBase`; Add the essential field `vision` to `PartyDTO`; Write the corresponding data to the result of `PartyService.getParty`.

+++ src/main/java/de/mfberg/bbak/controllers/AdminController.java
@@ -21,8 +21,8 @@ public class AdminController {
 
     @GetMapping("/worldmap")
     public ResponseEntity<List<HexTileDTO>> getWorldmap(
-            @RequestParam(defaultValue = "0") Integer aroundAxialQ,
-            @RequestParam(defaultValue = "0") Integer aroundAxialR,
+            @RequestParam(defaultValue = "0") long aroundAxialQ,
+            @RequestParam(defaultValue = "0") long aroundAxialR,
             @RequestParam(defaultValue = "3") byte radius
     ) {
             List<HexTileDTO> worldMap = worldAdminService.getHexTileDTOs(aroundAxialQ, aroundAxialR, radius);
